Pass configuration as analytics
