+++ snn_hfo_ieeg/entrypoint/argument_parsing.py
@@ -1,6 +1,6 @@
 import argparse
 import sys
-from snn_hfo_ieeg.stages.shared_config import Configuration, MeasurementMode, PlotMode
+from snn_hfo_ieeg.user_facing_data import Configuration, MeasurementMode, PlotMode
 from snn_hfo_ieeg.entrypoint.hfo_detection import CustomOverrides
 from snn_hfo_ieeg.plotting.plot_loader import find_plotting_functions
