+++ tests/integration/test_hfo_detection_from_dataset.py
@@ -1,7 +1,7 @@
 import pytest
 from snn_hfo_ieeg.plotting.persistence import PlotMode
 from snn_hfo_ieeg.user_facing_data import HfoDetection, Periods, Analytics, HfoDetectionWithAnalytics
-from snn_hfo_ieeg.stages.shared_config import Configuration, MeasurementMode
+from snn_hfo_ieeg.user_facing_data import Configuration, MeasurementMode
 from snn_hfo_ieeg.entrypoint.hfo_detection import run_hfo_detection_with_configuration
 from snn_hfo_ieeg.plotting.plot_loader import PlottingFunctions
 from tests.utility import assert_are_hfo_detections_equal
