+++ tests/integration/test_plotting.py
@@ -2,7 +2,7 @@
 import pytest
 from snn_hfo_ieeg.plotting.persistence import PlotMode
 from snn_hfo_ieeg.plotting.plot_loader import find_plotting_functions
-from snn_hfo_ieeg.stages.shared_config import Configuration, MeasurementMode
+from snn_hfo_ieeg.user_facing_data import Configuration, MeasurementMode
 from snn_hfo_ieeg.entrypoint.hfo_detection import run_hfo_detection_with_configuration
 from snn_hfo_ieeg.plotting.plot_channel import ChannelDebugError
 from snn_hfo_ieeg.plotting.plot_patient import PatientDebugError
